Modal: update api
